scripts/requirements.txt
google-generativeai>=0.3.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
requests>=2.31.0
python-dotenv>=1.0.0

